fix design-system page components
