+++ src/components/design-system/Logo.tsx
@@ -1,0 +1,51 @@
+import { motion } from "framer-motion"
+
+export const Logo = ({ isOpen }: { isOpen: boolean }) => {
+    return (
+      <a className="relative z-20 flex items-center space-x-2 py-1 text-sm font-normal" href="#">
+        <div className="h-8 w-8 shrink-0">
+          <svg
+            fill="none"
+            height="26"
+            viewBox="0 0 348 344"
+            width="26"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              className="fill-primary"
+              d="M334.974 109.247C317.709 113.038 301.421 119.405 286.525 127.904C291.885 141.586 294.876 156.453 294.876 172.001C294.876 230.105 253.86 278.614 199.22 290.223C200.464 308.94 204.64 326.856 211.333 343.5C289.22 326.412 347.531 257.054 347.531 174.044C347.531 151.122 343.059 129.267 334.974 109.247Z"
+            />
+            <path
+              className="fill-primary"
+              d="M294.869 172.001C294.869 156.423 291.878 141.557 286.518 127.904C234.129 157.756 198.799 214.083 198.799 278.703C198.799 282.582 198.977 286.403 199.213 290.193C253.853 278.584 294.869 230.075 294.869 171.971V172.001Z"
+            />
+            <path
+              className="fill-primary"
+              d="M53.1244 172.001C53.1244 156.423 56.1154 141.557 61.4757 127.904C46.5499 119.405 30.2617 113.038 13.0259 109.247C4.94107 129.267 0.469238 151.122 0.469238 174.044C0.469238 257.054 58.7808 326.412 136.668 343.5C143.361 326.856 147.566 308.94 148.78 290.223C94.1409 278.614 53.1244 230.105 53.1244 172.001Z"
+            />
+            <path
+              className="fill-primary"
+              d="M61.4852 127.874C56.1249 141.556 53.1338 156.423 53.1338 171.971C53.1338 230.075 94.1503 278.584 148.79 290.193C149.027 286.373 149.204 282.552 149.204 278.702C149.204 214.083 113.874 157.756 61.4852 127.904V127.874Z"
+            />
+            <path
+              className="fill-primary"
+              d="M262.38 89.6109C277.602 81.3188 291.432 70.8352 303.455 58.545C271.679 22.9184 225.48 0.5 174.009 0.5C125.441 0.5 81.5516 20.49 50.0415 52.6517C61.5024 65.4749 74.8291 76.5508 89.6069 85.5241C111.403 64.2607 141.166 51.1413 174.009 51.1413C208.895 51.1413 240.346 65.9487 262.409 89.6109H262.38Z"
+            />
+            <path
+              className="fill-primary"
+              d="M262.378 89.6111C240.315 65.9489 208.893 51.1415 173.977 51.1415C141.135 51.1415 111.342 64.2608 89.5752 85.5243C115.814 101.457 146.554 110.667 179.486 110.667C209.515 110.667 237.738 103.027 262.378 89.5815V89.6111Z"
+            />
+          </svg>
+        </div>
+        {isOpen && (
+          <motion.span
+            animate={{ opacity: 1 }}
+            className="font-medium whitespace-pre text-on-page"
+            initial={{ opacity: 0 }}
+          >
+            Theme Config
+          </motion.span>
+        )}
+      </a>
+    )
+  }
